refactor: split AthleteMapper into AthletePersistenceMapper and AthleteWebMapper (#50)
