+++ src/main/java/com/axel/trainingmetricsapi/controller/AthleteWebMapper.java
@@ -1,0 +1,30 @@
+package com.axel.trainingmetricsapi.controller;
+
+import com.axel.trainingmetricsapi.domain.Athlete;
+import com.axel.trainingmetricsapi.dto.request.AthleteRequest;
+import com.axel.trainingmetricsapi.dto.response.AthleteResponse;
+import org.springframework.stereotype.Component;
+
+@Component
+public class AthleteWebMapper {
+
+    public AthleteResponse domainToResponse(Athlete athlete){
+        return new AthleteResponse(
+            athlete.getId(),
+            athlete.getFirstName(),
+            athlete.getLastName(),
+            athlete.getBirthDate(),
+            athlete.getSport(),
+            athlete.getWeightInKg());
+    }
+
+    public Athlete requestToDomain(AthleteRequest athleteRequest) {
+        return new Athlete(
+            athleteRequest.firstName(),
+            athleteRequest.lastName(),
+            athleteRequest.birthDate(),
+            athleteRequest.sport(),
+            athleteRequest.weightInKg());
+    }
+    
+}
