+++ src/test/java/com/axel/trainingmetricsapi/repository/AthletePersistenceMapperTest.java
@@ -1,0 +1,65 @@
+package com.axel.trainingmetricsapi.repository;
+
+import com.axel.trainingmetricsapi.domain.Athlete;
+import com.axel.trainingmetricsapi.domain.Sport;
+import org.instancio.Instancio;
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class AthletePersistenceMapperTest {
+
+    private final AthletePersistenceMapper athletePersistenceMapper = new AthletePersistenceMapper();
+
+    @Test
+    void domainToEntity_shouldMapAllFields() {
+        Athlete athlete = Instancio.create(Athlete.class);
+        athlete.setId(null); // id null before persistence
+
+        AthleteJpaEntity athleteEntity = athletePersistenceMapper.domainToEntity(athlete);
+
+        assertAthleteFieldsMap(athlete, athleteEntity);
+        assertThat(athleteEntity.getId()).isNull();
+    }
+
+    @Test
+    void domainToEntity_shouldHandleNullableFields() {
+        Athlete athlete = new Athlete("Jean", "Dupont", null, Sport.TRIATHLON, null);
+
+        AthleteJpaEntity athleteEntity = athletePersistenceMapper.domainToEntity(athlete);
+
+        assertThat(athleteEntity.getBirthDate()).isNull();
+        assertThat(athleteEntity.getWeightInKg()).isNull();
+    }
+
+    @Test
+    void entityToDomain_shouldMapAllFields() {
+        AthleteJpaEntity athleteEntity = Instancio.create(AthleteJpaEntity.class);
+        athleteEntity.setId(42L);
+
+        Athlete athlete = athletePersistenceMapper.entityToDomain(athleteEntity);
+
+        assertAthleteFieldsMap(athlete, athleteEntity);
+        assertThat(athlete.getId()).isNotNull();
+    }
+
+    @Test
+    void entityToDomain_shouldHandleNullableFields() {
+        AthleteJpaEntity athleteEntity = new AthleteJpaEntity(2L, "Jean", "Dupont", null, Sport.TRIATHLON, null);
+
+        Athlete athlete = athletePersistenceMapper.entityToDomain(athleteEntity);
+
+        assertThat(athlete.getBirthDate()).isNull();
+        assertThat(athlete.getWeightInKg()).isNull();
+    }
+
+    private void assertAthleteFieldsMap(Athlete athlete, AthleteJpaEntity athleteEntity) {
+        assertThat(athlete.getFirstName()).isEqualTo(athleteEntity.getFirstName());
+        assertThat(athlete.getLastName()).isEqualTo(athleteEntity.getLastName());
+        assertThat(athlete.getBirthDate()).isEqualTo(athleteEntity.getBirthDate());
+        assertThat(athlete.getSport()).isEqualTo(athleteEntity.getSport());
+        assertThat(athlete.getWeightInKg()).isEqualTo(athleteEntity.getWeightInKg());
+        assertThat(athlete.getId()).isEqualTo(athleteEntity.getId());
+    }
+
+}
